Refactor: move file save to util function
