+++ src/components/ComponentSettingsModal.jsx
@@ -7,13 +7,10 @@ import { Redirect } from 'react-router-dom';
 import FormField from './FormField';
 import {
   frontMatterParser,
-  concatFrontMatterMdBody,
-  enquoteString,
   dequoteString,
-  generateResourceFileName,
-  generateCollectionPageFileName,
   generatePageFileName,
   retrieveCollectionAndLinkFromPermalink,
+  saveFileAndRetrieveUrl,
 } from '../utils';
 import elementStyles from '../styles/isomer-cms/Elements.module.scss';
 import { validatePageSettings, validateResourceSettings } from '../utils/validators';
@@ -176,80 +173,28 @@ export default class ComponentSettingsModal extends Component {
         title, permalink, fileUrl, date, mdBody, sha, category, prevCategory, baseApiUrl, type, thirdNavTitle,
       } = this.state;
       const { fileName, isNewFile, siteName } = this.props;
+
+      const fileInfo = {
+        title,
+        permalink,
+        fileUrl,
+        date,
+        mdBody,
+        sha,
+        category,
+        prevCategory,
+        baseApiUrl,
+        type,
+        thirdNavTitle,
+        fileName,
+        isNewFile,
+        siteName,
+      }
       
-      let newFileName, frontMatter
-      if (type === "resource") {
-        newFileName = generateResourceFileName(title, date);
-        frontMatter = { title: enquoteString(title), date };
-      } else if (type === "page") {
-        frontMatter = thirdNavTitle
-          ? { title, permalink, third_nav_title: thirdNavTitle }
-          : { title, permalink };
-        if (category) {
-          const groupIdentifier = fileName.split('-')[0];
-          newFileName = generateCollectionPageFileName(title, groupIdentifier);
-        } else {
-          newFileName = generatePageFileName(title);
-        }
-      }
-
-      if (permalink) {
-        frontMatter.permalink = `/${category ? `${category}/` : ''}${permalink}`;
-      }
-      if (fileUrl) {
-        frontMatter.file_url = fileUrl;
-      }
-      let newBaseApiUrl
-      if (prevCategory) {
-        // baseApiUrl can be used as is because user cannot change categories
-        newBaseApiUrl = baseApiUrl
-      } else {
-        if (category) {
-          // User is adding file to category from main page
-          newBaseApiUrl = `${baseApiUrl}/${type === "resource" ? `resources/${category}` : `collections/${category}`}`
-        } else {
-          // User is adding file with no collections, only occurs for pages
-          newBaseApiUrl = baseApiUrl
-        }
-      }
-
-      const content = concatFrontMatterMdBody(frontMatter, mdBody);
-      const base64EncodedContent = Base64.encode(content);
-
-      let params = {};
-      if (newFileName !== fileName || prevCategory !== category) {
-        // We'll need to create a new .md file with a new filename
-        params = {
-          content: base64EncodedContent,
-          pageName: newFileName,
-        };
-
-        // If it is an existing file, delete the existing page
-        if (!isNewFile) {
-          await axios.delete(`${newBaseApiUrl}/pages/${fileName}`, {
-            data: {
-              sha,
-            },
-          });
-        }
-        await axios.post(`${newBaseApiUrl}/pages`, params);
-      } else {
-        // Save to existing .md file
-        params = {
-          content: base64EncodedContent,
-          sha,
-        };
-        await axios.post(`${newBaseApiUrl}/pages/${fileName}`, params);
-      }
+      const newPageUrl = await saveFileAndRetrieveUrl(fileInfo)
+      console.log(newPageUrl)
       // Refresh page
       !isNewFile && window.location.reload();
-      
-      let newPageUrl
-      if (type === 'resource') {
-        newPageUrl = `/sites/${siteName}/resources/${category}/${newFileName}`
-      } else if (type === 'page') {
-        newPageUrl = category ? `/sites/${siteName}/collections/${category}/${newFileName}` : `/sites/${siteName}/pages/${newFileName}`
-      }
       this.setState({ redirectToNewPage: true, newPageUrl })
     } catch (err) {
       console.log(err);
